Fix jmp target for negative displacement

+++ src/remote/process.rs
@@ -179,7 +179,7 @@ impl Process {
     ) -> Result<usize> {
         let displacement = self.read_memory::<i32>(address + offset.unwrap_or(0x1))?;
 
-        Ok((address + length.unwrap_or(0x5)) + displacement as usize)
+        Ok(((address + length.unwrap_or(0x5)) as isize + displacement as isize) as usize)
     }
 
     pub fn resolve_rip(
@@ -190,7 +190,7 @@ impl Process {
     ) -> Result<usize> {
         let displacement = self.read_memory::<i32>(address + offset.unwrap_or(0x3))?;
 
-        Ok((address + length.unwrap_or(0x7)) + displacement as usize)
+        Ok(((address + length.unwrap_or(0x7)) as isize + displacement as isize) as usize)
     }
 
     fn get_process_id_by_name(process_name: &str) -> Result<u32> {
